경북대 BE_도기헌 5주차 과제 (1단계, 2단계) (#202)
* chore: 4주차 코드 옮겨오기

* add: order schema 추가

* add: gson, simple json dependancy 추가

* fix: url : wishlist->wishlists 변경

* feat: 카카오 로그인 및 마이페이지 추가

* feat: 카카오 로그인 및 마이페이지 뷰 컨트롤러 추가

* chore: categoryId->category_id

* feat: order 연관관계 추가

* feat: CONFLICT handler 추가

* feat: findby옵션 추가

* feat: existby옵션 추가

* add: amount 양수 조건 추가

* feat: 카카오 로그인

* chore: 반환형 object->map으로 변경

* delete: email exist 조건 제거

* feat: kakao login button, page 추가

* feat: 마이페이지 추가

* feat: order 기능 추가

* feat: order entity 추가

* feat: KakaoProperties 받아올 entity 추가

* test: OrderServiceIntegrationTest 추가

* test: findByProductIdAndOptionId 테스트 추가

* add: application.properties

* docs: 전체적으로 리펙터링 할거 체크리스트 추가

* docs: BCrypt 사용을 위해 dependancy 추가

* feat: 모든 route에 대해 filter 추가

* feat: admin 페이지용 filter 추가

* add: 모든 route에 대해 필터 추가

* add: 모든 엔티티에 user 연관관계 추가

* feat: authorization 체크 기능

* feat: user과의 연관관계 추가

* feat: findby 방식 추가

* feat: authorization-email 체크 기능

* add: user role 추가

* test: 테스트 정상화

* chore: formatting 변경, 따옴표 추가, 제목 변경

* chore: reformat code

* feat: response dtos 추가

* feat: 서버로부터 파라미터 받아와서 사용하도록 변경

* add: 카톡 나에게 보내기 위치 변경(controller->service)

* test: 테스트 이름 구체화

* chore: userToken, adminToken 위치 변경

* add: mock session으로 테스트하도록 변경

* refactor: addUrlPatterns url 관리 방식 리펙터링

* refactor: Map 대신 DTO로 응답하도록 변경

* refactor: response 공통 부분 메서드화

+++ src/main/java/gift/service/OptionService.java
@@ -2,9 +2,14 @@ package gift.service;
 
 import gift.entity.Option;
 import gift.entity.OptionDTO;
+import gift.entity.User;
 import gift.exception.ResourceNotFoundException;
 import gift.repository.OptionRepository;
+import gift.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -12,14 +17,20 @@ import java.util.List;
 public class OptionService {
 
     private final OptionRepository optionRepository;
+    private final UserService userService;
+    private final UserRepository userRepository;
 
-    public OptionService(OptionRepository optionRepository) {
+    @Autowired
+    public OptionService(OptionRepository optionRepository, UserService userService, UserRepository userRepository) {
         this.optionRepository = optionRepository;
+        this.userService = userService;
+        this.userRepository = userRepository;
     }
 
-    public Option save(OptionDTO optionDTO) {
-        Option option = new Option();
-        option.setOptionDTO(optionDTO);
+    public Option save(OptionDTO optionDTO, String email) {
+        User user = userService.findOne(email);
+        Option option = new Option(optionDTO);
+        option.setUser(user);
         return optionRepository.save(option);
     }
 
@@ -32,22 +43,43 @@ public class OptionService {
         return optionRepository.findAll();
     }
 
-    public Option update(Long id, OptionDTO optionDTO) {
+    public Option update(Long id, OptionDTO optionDTO, String email) {
+        if (!optionMatchesUser(id, email)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         Option option = findById(id);
         option.setOptionDTO(optionDTO);
+
         return optionRepository.save(option);
     }
 
-    public void delete(Long id) {
+    public void delete(Long id, String email) {
+        if (!optionMatchesUser(id, email)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
+        Option option = findById(id);
+        option.setUser(null);
+
         optionRepository.deleteById(id);
     }
 
     public Option subtract(Long id, int amount) {
         Option option = findById(id);
+        if (amount <= 0) {
+            throw new IllegalArgumentException("Invalid quantity");
+        }
         if (option.getQuantity() < amount) {
             throw new IllegalArgumentException("Not enough quantity");
         }
         option.subtract(amount);
         return optionRepository.save(option);
+    }
+
+    public boolean optionMatchesUser(Long id, String email) {
+        User user = userService.findOne(email);
+        Option option = findById(id);
+        return user.getId() == option.getUser().getId();
     }
 }
